requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.13.3
PyYAML==6.0.2
pytest==8.3.3
httpx==0.28.1
alembic==1.11.1
numpy==1.26.4
python-jose[cryptography]==3.3.0
